pyproject.toml
[tool.poetry]
name = "ai-code-review"
version = "0.1.0"
description = "AI-powered code review integration with Jira"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
jira = "^3.5.1"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
gitpython = "^3.1.40"

[tool.poetry.group.dev.dependencies]
black = "^23.10.1"
isort = "^5.12.0"
flake8 = "^6.1.0"
pytest = "^7.4.3"
httpx = "^0.25.1"
pre-commit = "^3.5.0"
flake8-docstrings = "^1.7.0"
flake8-bugbear = "^23.9.1"
flake8-comprehensions = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi-line-output = 3
line-length = 88
include-trailing-comma = true
force-grid-wrap = 0
use-parentheses = true
ensure-newline-before-comments = true